Move models variable to config
